Rewrite details_page in TS
